calc plot x_min_max with plots method
